# docs/swagger.yaml
definitions:
  internal.ErrorResponse:
    properties:
      error:
        example: internal server error
        type: string
    type: object
  internal.HealthResponse:
    properties:
      message:
        example: API is healthy
        type: string
    type: object
  internal.Student:
    properties:
      age:
        example: 20
        type: integer
      email:
        example: john@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
    type: object
  internal.StudentCreateRequest:
    properties:
      age:
        example: 20
        type: integer
      email:
        example: john@example.com
        type: string
      name:
        example: John Doe
        type: string
    type: object
  internal.StudentUpdateRequest:
    properties:
      age:
        example: 21
        type: integer
      email:
        example: john.new@example.com
        type: string
      name:
        example: John Doe
        type: string
    type: object
  internal.User:
    properties:
      email:
        example: john@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
    type: object
  internal.UserCreateRequest:
    properties:
      email:
        example: john@example.com
        type: string
      name:
        example: John Doe
        type: string
    type: object
info:
  contact:
    email: pratikraj220011@gmail.com
    name: Pratik Raj
  description: This is a sample server for managing student records.
  title: Student API Documentation
  version: "1.0"
paths:
  /health:
    get:
      description: Returns the API health status.
      operationId: healthcheck
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.HealthResponse'
      summary: Health check
      tags:
      - Health
  /students:
    get:
      description: Returns all students
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal.Student'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
      summary: List students
      tags:
      - Students
    post:
      consumes:
      - application/json
      description: Creates a new student
      parameters:
      - description: Student payload
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/internal.StudentCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
      summary: Create a student
      tags:
      - Students
  /students/{id}:
    delete:
      description: Deletes a student by ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
      summary: Delete a student
      tags:
      - Students
    get:
      description: Retrieves a student by its ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
      summary: Get a student by ID
      tags:
      - Students
    put:
      consumes:
      - application/json
      description: Updates an existing student by ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: Student payload
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/internal.StudentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
      summary: Update a student
      tags:
      - Students
  /users:
    get:
      description: Returns all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new user
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
      summary: Create a user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Deletes a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
      summary: Delete a user
      tags:
      - Users
  /users/by-id:
    get:
      description: Retrieve a single user by ID using query parameter ?id=
      parameters:
      - description: User ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal.ErrorResponse'
      summary: Get user by ID
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"